Firmware/tests/requirements.txt
requests
pyserial
